Use usize for loop counter

+++ rust/src/main.rs
@@ -3,13 +3,13 @@ use std::time::Instant;
 
 fn main() {
     let args: Vec<String> = env::args().collect();
-    let n: i64 = if args.len() > 1 {
+    let n: usize = if args.len() > 1 {
         args[1].parse().unwrap_or(10_000_000)
     } else {
         10_000_000
     };
     let start = Instant::now();
-    let mut sum: i64 = 0;
+    let mut sum: usize = 0;
     for i in 0..n {
         sum += i;
     }
